add adciidoc execution instructions
I got stuck trying to use this extension, as I didn't realise using it was as easy as adding a parameter when executing it.
I got stuck following the examples in this document:
http://asciidoctor.org/news/2014/02/18/plain-text-diagrams-in-asciidoctor/
which was a lot more complicated and failed on asciidoctor 1.5.2
also, it was confusing that this doc referred to asciidoctor-diagrams as an "extension", but in the command line tool's help, the -r parameter added a "library" ... I wasn't sure this was the same thing

Thanks for this really nice tool!
Keep it up!